requirements.txt
flask
dash
dash_bootstrap_components
pandas
plotly
sqlalchemy
psycopg2-binary
requests
openpyxl
pdfplumber
python-docx
opencv-python
ultralytics
